// src/recipe/thirdpartyemailpassword/types.ts
export interface User {
  id: string
  timeJoined: number
  email: string
  thirdParty?: {
    id: string
    userId: string
  }
}
// ...
export interface RecipeInterface {
  getUserById(input: { userId: string; userContext: any }): Promise<User | undefined>

  getUsersByEmail(input: { email: string; userContext: any }): Promise<User[]>

  getUserByThirdPartyInfo(input: {
    thirdPartyId: string
    thirdPartyUserId: string
    userContext: any
  }): Promise<User | undefined>

  thirdPartySignInUp(input: {
    thirdPartyId: string
    thirdPartyUserId: string
    email: string
    userContext: any
  }): Promise<{ status: 'OK'; createdNewUser: boolean; user: User }>

  emailPasswordSignUp(input: {
    email: string
    password: string
    userContext: any
  }): Promise<{ status: 'OK'; user: User } | { status: 'EMAIL_ALREADY_EXISTS_ERROR' }>

  emailPasswordSignIn(input: {
    email: string
    password: string
    userContext: any
  }): Promise<{ status: 'OK'; user: User } | { status: 'WRONG_CREDENTIALS_ERROR' }>

  createResetPasswordToken(input: {
    userId: string
    userContext: any
  }): Promise<{ status: 'OK'; token: string } | { status: 'UNKNOWN_USER_ID_ERROR' }>

  resetPasswordUsingToken(input: {
    token: string
    newPassword: string
    userContext: any
  }): Promise<
        | {
          status: 'OK'
          /**
               * The id of the user whose password was reset.
               * Defined for Core versions 3.9 or later
               */
          userId?: string
        }
        | { status: 'RESET_PASSWORD_INVALID_TOKEN_ERROR' }
    >

  updateEmailOrPassword(input: {
    userId: string
    email?: string
    password?: string
    userContext: any
  }): Promise<{ status: 'OK' | 'UNKNOWN_USER_ID_ERROR' | 'EMAIL_ALREADY_EXISTS_ERROR' }>

}